requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==3.2.5
django-on-heroku==1.1.2
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
